Add handling of composition of standard types
Current implementation does not allow to:
* put variant, pair, tuple inside optional,
* put pair, tuple inside variant,
* put pair inside tuple

This is just because order of definitions of to_string specialisations
matters. When compiler matches to_string for optional it is not
aware of existence of specialisations for variant, pair, tuple.

This change propose to add forward declarations to to_string
specialisations for variant, pair and tuple.