Require whitespace before binary `*`, `&`, `~`, closes #152
Before this I had been trying out making `a*b` work without whitespace by making `a****b` interpret the final `*` as binary if followed by something that wouldn't make sense if it was unary, but that's a bit subtle. (And similarly for `&` and should also include `~`)

See additional discussion in #152.