Treat `operator` and operator symbol as a single identifier token
For example, `operator+ : (a: mytype, b:mytype) = { /*... */}` makes `operator +` a single identifier token.
That seemed to be the simplest and most elegant way to have the rest of the compiler downstream just treat it as a single identifier.
Do a pass tidying up the grammar, particularly to make keywords/operators distinct now that there's an _operator_ production and an `operator` keyword.