Add return forward passing style support
Based on chapter `1.5 Return values, D0708 R0` cppfront
should support forward return passing style. The current
implementeation makes it impossible to pass variable as a reference,
or to forward value with preserving the return-expression’s
cv-qualification and value category.

After this change it is possible to create the following function:
```cpp
first: (forward rng) -> forward _ = {
    [[assert Bounds: !std::empty(rng)]]
    return std::begin(rng)*;
}
```
that will (for lvalue rng) returns lvalue reference to the first element
of the rng and will make the following code possible:
```cpp
main: () -> int = {
    v : std::vector = (1,2,3);
    first(v) = 4;
    std::cout << first(v) << std::endl; // prints: 4
}
```

Limitations:

* don't know if `passing_style::move` has a special way
or is it current default one.
* the wildcar is used currently `forward _` to get `decltype(auto)`,
* wildcars make these function order dependent and need to be defined
before it is used.