Require that floating-point literals write a digit after `.`, enables UFCS on numeric literals
(1) I find `1.0 and `1.0f` clearer than `1.` and `1.f` anyway, so that's a small plus.
(2) This was the only thing in the way of using UFCS on numeric literals, so that's another maybe-medium plus.

And the latter enables writing unit libraries like `42.gram()`, as proposed in #284's comment thread, which is an interesting use of UFCS on literals. So I'll mark this as closes #284.