Fix string literal parsing when \\ is on the end
The issue is that when there is '\\' on the end of the string
`expand_string_literal()` function goes out of string range
because for loop has a check
```cpp
pos < length && (text[pos] != '"' || text[pos-1] == '\\')
```
that is true when '\\' is before last '"' character that cause
loop to go beyond it and cause assertion to fail.

The current change adds a check if the '\' character before last '"'
is not escaped by the former '\' character.