Added type and namespace alias declarations, closes #273
An alias can be to a namespace, type, function, or object.

Aliases are always a non-mutable view of the original entity. This is natural for namespace and type and function aliases, since those kinds of entited can't be mutated, but it bears calling out for an object alias which behaves as `auto const&`.

The current syntax is like non-alias declarations, but with `==` to connote equality with an existing entity, instead of `=` which connotes setting the new entity's value. For example:

    lit: namespace == ::std::literals;

    pmr_vec: <T> type
        == std::vector<T, std::pmr::polymorphic_allocator<T>>;

    func :== some_original::inconvenient::function_name;

    vec :== my_vector;  // note: const&, aliases are never mutable

Note: Function aliases are subsumed under the object case, so you can't declare a function alias with an explicit signature. Rationale:
    - if there's no need for a (repetitive) explicit signature,
        the object case covers it
    - if the parameters/returns must be exact (no conversions),
        a pointer to function covers it
    - if parameter/return conversions are allowed,
        std::function covers it
But if we do learn about a reason to add support for function aliases having an explicit signature, and possibly with conversions to/from the parameter/return types, they'll be a natural fit here.