Start moving more semantic checks into `sema`
Longer version:

Until now, I've been doing some semantic checking early even during parsing (on the grounds of ‘hey why not catch mistakes early?’), some in `sema.h` (definite first/last use calculation), and some in `cppfront.cpp` during lowering.

As part of yesterday’s initial checkin for reflection and code generation, I started moving some checks from `parse.h` to later. This is because when generating code like the `interface` virtual destructor, first we lex/parse and then we add the new member into the type being defined, but we don't want the check to happed immediately when the generated code is parsed in isolation, we want to wait for that next step to add it to the type. So that answered my rhetorical question above of 'hey why not catch mistakes early' because the observation is that, in the presence of the ability for a meta function to generate code that will be part of the definition of the type and so effectively participates in the final phase of parsing a type, we want that code generation to finish adding things to the type and then let semantic checking see the whole result together.

So I'll gradually be moving semantic rule checking code from `parse.h` and `cppfront.h` into `sema.h`, and this commit is part of that for some of the checks I just moved to `cppfront.cpp` yesterday, so that they live in `sema.h` instead.