Added a second meta function from P0707, polymorphic_base
And, dang it, wouldn't you know that this found yet another (and the same) bug in my test code - another virtual destructor I forgot to define in the test example, this time in `Machine`. Now the meta function does it for me. Sweet.

> Naming note: P0707 calls this meta function `base_class`, but since I'm using the word "type" instead of "class" in Cpp2, and also because not all bases are necessarily polymorphic (see `cpp2::meta` for counterexamples), I think I like this name better.

(Also, moved another batch of the sema checks into `sema`.)

I did notice one grammar detail... currently the meta function name comes first, which means that a short-form `<T>` template parameter list can look syntactically like a template argument list. That is,

```
X: @mymetafunc <T> type = ...
```

parses as having the meta function name `mymetafunc<T>`, because of max munch. That's okay, if you wanted the `<T>` to be a template parameter list, you can just write it out as `<T: type>` and it's fine. But this might be a reason to put the meta function name first, like

```
@mymetafunc X: <T> type = ...
```

which those used to Python style (or who just prefer a newline) will write as

```
@mymetafunc
X: <T> type = ...
```

... aaaaand we'll forever be accused of slavishly copying Python. :grin: (IIUC, Python had different reasons to put the decorator before the name.)

For now I'll leave the ~decorator~ meta function name where it is, because I kind of like the "adjective `type`" naming style (ever since I did `interface class` for C++/CLI), but this is something that's easy to tweak based on experience and feedback...