Fix 'operator=` memberwise generation from `that`
The `operator=` memberwise generation code didn't correctly handle the case of not-explicitly-initialized members where there are no more statements in the function (including all of them in an empty function).

And it was a silly mistake -- the main loop was over all objects *and statements*, and ended when we ran out of statements -- and then there was a whole 'nother separate loop for the rest of the objects not initialized yet. When you say something twice, you get it wrong the second time... the separate loop wasn't doing all the work of the main loop.

The fix is to jettison that separate loop, make the main loop handle all objects (only) and have a conditional guard for the case where we've run out of explicit statements to try first as the preferred explicit value-set candidate.