Add `alien_memory<T>` aka `'T volatile`, see #393 comment thread
The main problem with `volatile` isn't the semantics -- those are are deliberately underspecified, and appropriate for talking about "memory that's outside the C++ program that the compiler can't assume it knows anything about" which is an important low-level concept. The problems with `volatile` are that (a) it's wired into the language as a type qualifier which is undesirable and unnecessary, and (b) the current name is confusing and has baggage and so it should be named something that connotes what it's actually for (and I like "alien" rather than "foreign" because I think "alien" has a better and stronger connotation).