Make output `-Wunused-parameter`-clean, and add real support for `_` unnamed parameters, closes #560
Emit `[[maybe_unused]]` on a `that` parameter if the class is empty.
Add -Wunused-parameter to my GCC and Clang regression tests.
Add `_` unnamed parameter support, and use ordinals instead of line/col numbers to keep the generated names stable.

As an example from one of the updated regression tests that had a helper function to just accept (but not use) five generic parameters, that function is now just: `call: (_, _, _, _, _) = { }`