Restore GCC 10 test results, and add diagnostic for the one construct that's known not to work on GCC 10
Although I'd previously increased the minimum level of GCC support from 10.x to 11.x, in practice most things work just fine. The only known exception is a `forward` parameter of a non-wildcard type on a type-scope (aka member) function.

So this commit:
- restores the `test-results/gcc-10` directory (while still keeping the `gcc-13` one too)
- adds a check for the known case that doesn't work in GCC 10

I don't intend to do a lot of work to support downlevel compilers, but if it's just one rarely-used item then it's worth trying to maintain downlevel compatibility. If we encounter other weaknesses with GCC 10, I'll likely just re-raise the minimum GCC version back to 11.