Add partial support for `@union`
A `@union type` is a "language variant," but expressed as a type metafunction rather than specially baked into the language.

This is only very basic support so far, much more is needed including `...` variadic construction and `<<` streaming, both of which need Cpp2 language upgrades to support `...` variadic parameters and emission of `friend` functions for streaming operators.

Also in this commit:

- For statement parameters, change the default passing style to `copy` (experiment)

- Support `that` on all type-scope functions, even if they don't have `this`

- Support `as std::string` for the same cases that work for string interpolation... especially int to string. Yes, `std::to_string(myint)` is already better than we had before, but `value as std::string` just keeps rolling off my fingertips and it's time for it to work