Let metafunctions replace specific members, and apply to `basic_enum`
Before, `basic_enum` was just removing all the statements in a type's body, not just the object declarations it was replacing. This  has two problems:
- It masked programmer errors that would normally have been diagnosed in sema, such as writing an assignment statement by accident (e.g., `wrong: @enum type = { oops = assignment; /* intended to write := here, but the errant statement was just silently removed */ }`
- It disabled the ability for programmers to do things like write member functions of enumeration types

Now that `basic_enum` removes only the data members that it processes, any other code inside the body will be diagnosed or allowed, as appropriate.