Update `[[nodiscard]]` default for streaming to match established convention
We already do this for assignment: As with assignment, the convention for streaming operators is to return a reference to the stream, to permit chaining. But ignoring that return value at the end of each chain is always expected and always correct, and so those return values should always be discardable.