Re-enable move from last use in initializers, closes #749
I was trying out suppressing move-from-last-use if the last use was in an initializer, because of examples like `view: std::span = local_variable;` when that was the last use of `local_variable`, which is an error for an rvalue. But that also suppresses a righteous move in examples like `result := local_variable;`. So I now agree those should be treated as rvalues too, and if that causes an error for an rvalue because there is no further us of that variable in scope, then the programmer must write an explicit discard of the variable.