Diagnose named-but-unused local variables
This is a case of losing a non-discarded value that wasn't being covered with the discarding rule.

I think I like this... although I had to update unused dummy variables in the tests to rename them to `_`, it caught a few cases where I had a loop `counter` variable I wasn't actually using that was just a leftover from a previous incarnation of the test, so I could and did remove it.