Enable natural use of single named return value
Given:

    f: () -> (ret: int = 42)
    = {
        // ...
    }

The call site `f().ret` has always worked, as it should for a named return.

The call site `f()` now also works -- we can omit the name if we don't need it. This lets callers have the same convenience of a single unnamed return value, while giving function authors the convenience of writing a named return value, including so the return value has a name for postconditions:

    f: () -> (ret: int = 42)
        post (ret > 0)
    = {
        // ...
    }