On second thought, just remove the one use of `ranges::`
There's only one use of `ranges::` in the codebase, and I want cppfront to compile on reasonably recent compilers. Since Apple Clang 14 only started supporting ranges in March 2023, and that seems to be the only thing stopping use of Clang 12+ even on Xcode (AFAIK), it seems more reasonable to do this and not break Xcode Clang 12+ builds of cppfront.