Match C++ standard used in compiler-specific run-tests scripts to the CI script
f0018d757d09aae09d18a25a6b22bc3e24eecf85 modified the set C++ standard for some compilers in the CI script, but did not do so for the compiler-specific run-tests scripts. This fixes the discrepancy by matching setting in the compiler-specific scripts to that of the CI script (since the above commit apparently did update the tests results according to the newly set standard in the CI script).

(It would be more robust to avoid the possibility of such a discrepancy by avoiding having such settings in multiple scripts.)