Create pure2-regex-partial.cpp2
Feel free to drop this file again:

I just added it to show the results of trying to comment out all the cases that cause the metafunction to report an error, in an attempt to get an executable file. After commenting these ones, cppfront completes

I still get strange errors from the Cpp1 compiler, which I've narrowed down to this short repro:

    #define CPP2_IMPORT_STD          Yes
    #include "cpp2util.h"

If the macro is commented out, things compile fine. So there's something about the "import std" path that's going wrong, it seems on both MSVC and GCC 10. I haven't been able to diagnose the problem further than that though.