Extend is() to correctly handle custom is operator
In current implementation custom is operator (`is_op()` member function)
suffers from implicit casting of its argument. The same issue that were
with function predicates.

After this change we added the check if the cast is lossless or
if the custom is operator is generic function.

Added tests to verify is with various values.

Added new concepts to make checks more readable:
- valid_custom_is_operator
- defined - to check if type is defined (used to determined generic fun)
- predicate_member_fun - to check if member function is a predicate,

Added new type trait:
- argument_of_op_is_t to determined the type of op_is argument,

Correct helper methods:
- argument_of_helper - to handle ref-qualified member functions,
- argument_of_helper - to work with non-copyable types,