Made range operators explicit about whether the last value is included
The original design was to support `...` (implicitly half-open, last value is excluded) and `..=` (explicitly closed, last value is included). This changes that to `..<` (explicitly half-open, last value is excluded) and `..=` (explicitly closed, last value is included).

Rationale: Originally I made `...` the default because it's the safe default, for common cases like iterator ranges where including the last element is a bounds violation error. It would be creating a terrible pitfall to make the "default" syntax be an out-of-bounds error on every use with iterators.

However, feedback on Reddit and elsewhere quickly pointed out that for numeric ranges `...` not including the last value is also surprising. What to do?

One way out is to not support iterators. However, that would be a needless loss of functionality if there was a better answer.

And there is: A better way out is to simply embrace that there should not be a default, but make it  convenient for programmers to be explicit every time about whether the last element is included or not. Supporting `..<` and `..=` as the range operators achieves that goal.

I appreciate the feedback, and I think it led to a superior design for a C++-compatible ecosystem that heavily uses iterators today. Thanks to everyone who gave feedback!